libultra warnings fixed (#336)
* Separated libultra into it's own directory and built it as a static library

* Added first decompiled file (osSetTime) and set up makefile to correctly compile it, modified linker script to put it in the right place

* Added sprintf, osMapLB/UnmapTLBAll

* added more decomps, removed references to libultra.h outside of ultra64.h

* osCreateViManager with all it's data safely stored in the same file

* osCreatePiManager done

* US ok now, overrode osCreateViManager.o's text alignment in the linker file

* osSetThreadPri done, both OK

* osInitialize, both OK

* added osViSwapBuffer and cleaned up old code, removed int in favor of u32/s32 and renamed the __osDisableInt variable from res to int_disabled

* actually added osViSwapBuffer to sm64.ld and removed from libultra.s

* sqrtf both OK

* osContStartReadData both OK

* osContInit both OK

* osEepromProbe both OK

* osEepromProbe both OK

* libultra_math, osInvalDCache.  added tool to patch libultra_math object file so it links.  both OK

* osPiStartDma both OK

* bzero and osMapTLB both OK

* osEepromLongRead both OK

* osEepromLongWrite both OK

* bcopy both OK

* gu is O3, guOrthoF done both OK

* guPerspective done both OK i hate rodata

* osGetTime done both OK

* llconv done both OK

* skipping sinfcosf for now...

* guTranslateF done both OK

* sinf and cosf both OK!

* cleaned up libultra code and did some minor documenting

* guTranslateF is actually O3.  fixed

* guRotate finally both OK, thanks simon!

* guScale done both OK

* osAiSetFrequency done both OK

* alBnkfNew both OK thanks simon, should probably figure out bss reorderings at some point

* osWritebacakDCache both OK

* osAiGetLength done both OK

* osAiSetNextBuffer done and documented a bit

* osTimer done both OK

* _Printf done both OK

* string done both OK

* dequeue thread, disable/restore int both OK

* __osViInit done.  a lot of rearranging to get the data right.  both OK

* osexception both OK

* osvirtualtophysical done both ok

* osspsetstatus done both OK

* osSpSetPc done both OK

* osSpRawStartDma done both OK

* osSpDeviceBusy done both OK

* osSpGetStatus done both OK

* osGetThreadPri and osViGetCurrentContext done both OK

* osViSwapContext done both OK

* osGetCount done

* osPiCreateAccessQueue done both OK

* moved some rodata around

* added 2 more functions to picreateaccessqueue oops

* osPiRawStartDMA and also created the osViModeTable..

* osDevMgrMain done both OK

* some asm functions done

* osSiRawRead/WriteIo done both OK

* osMapTLBRdb

* osPiRawReadIo done

* __osSiCreateAccessQueue done both OK

* osSiRawStartDma done both OK

* osSetTimer done both OK

* osEepromWrite plus some documenting

* osJamMesg done

* osPiGetCmdQueue plus some documentation

* osEepromRead done both OK

* guMtxF2L done both OK... finally *(s32*)&

* guNormalize done both OK

* osAiDeviceBusy and osSetCompare done both OK

* _Litob done both OK

* _Ldtob done both OK

* kdebugserver done both OK

* __osSyncPutChars done, both OK, bss and data for libultra done

* osSetIntMask done both OK

* osDestroyThread done both OK

* osProbeTlb and osSiDeviceBusy done both OK

* getCause and atomicDec done both OK

* minor cleanup

* finishing merge

* Hacky fix for reorderings

* fixed bss reorderings after libultra merge

* Small cleanups

* Better bss fixes

* dos2unix sm64.ld

* Whitespace fix

* Another whitespace fix

* addressing pr comments

* fixed all warnings

+++ lib/src/math/llmuldiv.c
@@ -1,46 +1,44 @@
-#include <ultra64.h>
-
-u64 __ull_rshift(u64 a0, u64 a1)
+unsigned long long __ull_rshift(unsigned long long a0, unsigned long long a1)
 {
     return a0 >> a1;
 }
-u64 __ull_rem(u64 a0, u64 a1)
+unsigned long long __ull_rem(unsigned long long a0, unsigned long long a1)
 {
     return a0 % a1;
 }
-u64 __ull_div(u64 a0, u64 a1)
+unsigned long long __ull_div(unsigned long long a0, unsigned long long a1)
 { 
     return a0 / a1;
 }
 
-u64 __ll_lshift(u64 a0, u64 a1)
+unsigned long long __ll_lshift(unsigned long long a0, unsigned long long a1)
 {
     return a0 << a1;
 }
 
-s64 __ll_rem(u64 a0, s64 a1)
+long long __ll_rem(unsigned long long a0, long long a1)
 {
     return a0 % a1;
 }
 
-s64 __ll_div(s64 a0, s64 a1)
+long long __ll_div(long long a0, long long a1)
 {
     return a0 / a1;
 }
 
-u64 __ll_mul(u64 a0, u64 a1)
+unsigned long long __ll_mul(unsigned long long a0, unsigned long long a1)
 {
     return a0 * a1;
 }
 
-void __ull_divremi(u64 *div, u64 *rem, u64 a2, u16 a3)
+void __ull_divremi(unsigned long long *div, unsigned long long *rem, unsigned long long a2, unsigned short a3)
 {
     *div = a2 / a3;
     *rem = a2 % a3; 
 } 
-s64 __ll_mod(s64 a0, s64 a1)
+long long __ll_mod(long long a0, long long a1)
 {
-    s64 tmp = a0 % a1;
+    long long tmp = a0 % a1;
     if ((tmp < 0 && a1 > 0) || (tmp > 0 && a1 < 0))
     {
 
@@ -49,7 +47,7 @@ s64 __ll_mod(s64 a0, s64 a1)
     return tmp;
 }
  
-s64 __ll_rshift(s64 a0, s64 a1)
+long long __ll_rshift(long long a0, long long a1)
 {
     return a0 >> a1;
 }
